requirements.txt
apache-airflow-providers-apache-kafka
kafka-python
psycopg2-binary